refactor(eval): move dispatch and the ceiling to the modules that own them
